Let BidirectionalSearch use different searches
