+++ core/src/main/java/aima/core/search/basic/support/BidirectionalSearch.java
@@ -1,13 +1,10 @@
-package aima.core.search.basic.uninformed;
+package aima.core.search.basic.support;
 
 import aima.core.search.api.BidirectionalProblem;
 import aima.core.search.api.Node;
 import aima.core.search.api.NodeFactory;
 import aima.core.search.api.Problem;
 import aima.core.search.api.SearchController;
-import aima.core.search.basic.support.BasicNodeFactory;
-import aima.core.search.basic.support.BasicSearchController;
-import aima.core.search.basic.support.StateActionTimeLine;
 
 import java.util.AbstractMap.SimpleEntry;
 import java.util.Comparator;
@@ -21,48 +18,48 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ForkJoinPool;
+import java.util.concurrent.PriorityBlockingQueue;
+import java.util.function.ToDoubleFunction;
 
 /**
- * find a solution by searching from a start and goals in parallel. When multiple solutions
- * exist, a random one is returned, see unit test arad2oreda. The processes expand states
- * parallel and check each others frontiers if they meet.
- * <p>
- * An optional {@link BiDirectionalBreadthFirstSearch#history} can be used for println debugging
- * but may slow down execution significantly
- *
- * @param <A> actions to take
- * @param <S> states which exist
+ * solves a {@link aima.core.search.api.BidirectionalProblem} with different {@link Strategy}
+ * @author wormi
  */
-public class BiDirectionalBreadthFirstSearch<A, S> {
+public class BidirectionalSearch<A, S> {
 
+  private final String forwardStrategy;
+  private final String backwardStrategy;
+  private final ToDoubleFunction<Node<A, S>> forwardHeuristicFunction;
+  private final ToDoubleFunction<Node<A, S>> backwardHeuristicFunction;
   private NodeFactory<A, S> nodeFactory = new BasicNodeFactory<>();
   private SearchController<A, S> searchController = new BasicSearchController<>();
-  private StateActionTimeLine<String, String> history;
-  private final double pathCost = 0;
 
-  /**
-   * @param biDirectionalProblem to solve
-   * @return list of actions to solve the problem
-   */
-  public List<A> findSolution(BidirectionalProblem<A, S> biDirectionalProblem) throws ExecutionException,
-      InterruptedException {
+  public BidirectionalSearch(String forwardStrategy, String backwardStrategy,
+                             ToDoubleFunction<Node<A, S>> forwardHeuristicFunction,
+                             ToDoubleFunction<Node<A, S>> backwardHeuristicFunction
+  ) {
+    this.forwardStrategy = forwardStrategy;
+    this.backwardStrategy = backwardStrategy;
+    this.forwardHeuristicFunction = forwardHeuristicFunction;
+    this.backwardHeuristicFunction = backwardHeuristicFunction;
+  }
 
+  public List<A> findSolution(BidirectionalProblem<A, S> bidirectionalProblem)
+      throws ExecutionException, InterruptedException {
     // build everything needed for the two searches
     Set<S> exploredFromStart = concurrentStateSet();
     Set<S> exploredFromGoal = concurrentStateSet();
-    Queue<Node<A, S>> fromStartFrontier = newFIFOQueue();
-    Queue<Node<A, S>> fromGoalFrontier = newFIFOQueue();
-    final Problem<A, S> originalProblem = biDirectionalProblem.getOriginalProblem();
-    final List<Problem<A, S>> reverseProblems = biDirectionalProblem.getReverseProblems();
+    Queue<Node<A, S>> fromStartFrontier = getQueue(forwardStrategy, forwardHeuristicFunction);
+    Queue<Node<A, S>> fromGoalFrontier = getQueue(backwardStrategy, backwardHeuristicFunction);
+    final Problem<A, S> originalProblem = bidirectionalProblem.getOriginalProblem();
+    final List<Problem<A, S>> reverseProblems = bidirectionalProblem.getReverseProblems();
 
     final S initialState = originalProblem.initialState();
-    Node<A, S> startNode = newRootNode(initialState, pathCost);
-    //rememberEvent("addToFrontier: " + startNode.state(), true);
+    Node<A, S> startNode = newRootNode(initialState, 0);
     fromStartFrontier.add(startNode);
     reverseProblems.forEach(problem -> {
       final S state = problem.initialState();
-      rememberEvent("addToFrontier: " + state, false);
-      fromGoalFrontier.add(newRootNode(state, pathCost));
+      fromGoalFrontier.add(newRootNode(state, 0));
     });
 
     // build Searches
@@ -82,8 +79,8 @@ public class BiDirectionalBreadthFirstSearch<A, S> {
           if (problem.isGoalState(nextNodeFromFrontier.state())) {
             return solution(nextNodeFromFrontier);
           }
-          // the whole work is done here, expand the node, add and remove frontier nodes and check if
-          // our parallel searches meet
+          // the whole work is done here, expand the node, add and remove frontier nodes and check
+          // if our parallel searches meet and calculate the resulting path
           while (nextNodeFromFrontier != null) {
             Node<A, S> nodeToExpand = nextNodeFromFrontier;
             Optional<SimpleEntry<Node<A, S>, Node<A, S>>> _solution =
@@ -92,11 +89,9 @@ public class BiDirectionalBreadthFirstSearch<A, S> {
               return solution(buildResultPath(_solution.get().getKey(), _solution.get().getValue(),
                   problem));
             }
-            rememberEvent("remove from frontier " + nodeToExpand.state(), searchDirection.forwardOrBackward);
             searchDirection.frontier.remove(nodeToExpand);
             nextNodeFromFrontier = searchDirection.frontier.peek();
           }
-          rememberEvent("abort search, frontier is empty", searchDirection.forwardOrBackward);
           return null;
         }).filter(result -> result != null).findFirst().orElse(failure());
 
@@ -105,63 +100,51 @@ public class BiDirectionalBreadthFirstSearch<A, S> {
 
     // execute search
     return forkJoinPool.submit(task).get();
+
   }
 
-
-  private Optional<SimpleEntry<Node<A, S>, Node<A, S>>> findMeetingNodes(Problem<A, S> problem,
-                                                                         SearchDirection
-      searchDirection, Node<A, S> nodeToExpand) {
-
-    Queue<Node<A, S>> frontier = searchDirection.frontier;
-    List<A> actions = problem.actions(nodeToExpand.state());
-    Set<S> exploredStates = searchDirection.exploredStates;
-    Queue<Node<A, S>> otherFrontier = searchDirection.otherFrontier;
-
-    rememberEvent("expand " + nodeToExpand.state(), searchDirection.forwardOrBackward);
-    return actions.parallelStream()
-        .map(action -> meetingNodes(exploredStates, frontier, otherFrontier,
-            newLeafNode(problem, nodeToExpand, action), searchDirection.forwardOrBackward))
-        .filter(o -> o != null)
-        .findFirst();
+  private Set<S> concurrentStateSet() {
+    return new ConcurrentSkipListSet<S>() {
+      @Override
+      public Comparator<? super S> comparator() {
+        return (Comparator<S>) (o1, o2) -> o1.toString().compareTo(o2.toString());
+      }
+    };
   }
 
-  private void rememberEvent(String action, boolean direction) {
-    if (history != null) {
-      rememberEvent(direction ? "forward" : "backward", action);
+  public Queue<Node<A, S>> getQueue(String strategy, ToDoubleFunction<Node<A, S>> h) {
+    if (Strategy.BFS.equals(strategy)) {
+      return newFIFOQueue();
     }
-  }
-
-  private void rememberEvent(String state, String action) {
-    history.add(state, action);
-  }
-
-  /**
-   * add a new node to the frontier if the state is not explored
-   * @return true when the new node reaches one of the other searchDirections branch
-   */
-  private SimpleEntry<Node<A, S>, Node<A, S>> meetingNodes(Set<S> exploredStates,
-                                                           Queue<Node<A, S>> thisFrontier,
-                                                           Queue<Node<A, S>> otherFrontier,
-                                                           Node<A, S> leaf,
-                                                           boolean direction) {
-    if (!exploredStates.contains(leaf.state())) {
-      rememberEvent("addToFrontier: " + leaf.state(), direction);
-      thisFrontier.add(leaf);
+    int initialCapacity = 10;
+    if (Strategy.UCS.equals(strategy)) {
+      return new PriorityBlockingQueue<>(initialCapacity, (o1, o2) -> Double.compare(o1.pathCost
+          (), o2.pathCost()));
     }
-    if (exploredStates.contains(leaf.state())) {
-      return null;
-    } else {
-      exploredStates.add(leaf.state());
+    if (Strategy.ASTAR.equals(strategy)) {
+      if (h == null) {
+        throw new IllegalArgumentException("A* needs a heuristic");
+      }
+      return new PriorityBlockingQueue<>(initialCapacity, (o1, o2) ->
+          Double.compare(o1.pathCost() + h.applyAsDouble(o1), o2.pathCost() + h.applyAsDouble(o2)));
     }
-    return meetingNodes(leaf, otherFrontier, direction);
+    throw new IllegalArgumentException("did not recognize strategy");
   }
 
-  /**
-   * @return node with equal state from the other frontier or null if none exists
-   */
+  public Node<A, S> newRootNode(S initialState, double pathCost) {
+    return nodeFactory.newRootNode(initialState, pathCost);
+  }
+
+  public List<A> failure() {
+    return searchController.failure();
+  }
+
+  public List<A> solution(Node<A, S> child) {
+    return searchController.solution(child);
+  }
+
   private Node<A, S> findNodeWithEqualState(Node<A, S> leaf, Queue<Node<A, S>> otherFrontier,
                                             boolean direction) {
-    rememberEvent("search for " + leaf.state() + "in other frontier", direction);
     return otherFrontier.parallelStream()
         .filter(node -> node.state().equals(leaf.state()))
         .findFirst()
@@ -209,8 +192,40 @@ public class BiDirectionalBreadthFirstSearch<A, S> {
     return result;
   }
 
-  public Node<A, S> newRootNode(S initialState, double pathCost) {
-    return nodeFactory.newRootNode(initialState, pathCost);
+  private Optional<SimpleEntry<Node<A, S>, Node<A, S>>> findMeetingNodes(Problem<A, S> problem,
+                                                                         SearchDirection searchDirection,
+                                                                         Node<A, S> nodeToExpand) {
+    Queue<Node<A, S>> frontier = searchDirection.frontier;
+    List<A> actions = problem.actions(nodeToExpand.state());
+    Set<S> exploredStates = searchDirection.exploredStates;
+    Queue<Node<A, S>> otherFrontier = searchDirection.otherFrontier;
+
+    return actions.parallelStream()
+        .map(action -> meetingNodes(exploredStates, frontier, otherFrontier,
+            newLeafNode(problem, nodeToExpand, action), searchDirection.forwardOrBackward))
+        .filter(o -> o != null)
+        .findFirst();
+  }
+
+  /**
+   * add a new node to the frontier if the state is not explored
+   *
+   * @return true when the new node reaches one of the other searchDirections branch
+   */
+  private SimpleEntry<Node<A, S>, Node<A, S>> meetingNodes(Set<S> exploredStates,
+                                                           Queue<Node<A, S>> thisFrontier,
+                                                           Queue<Node<A, S>> otherFrontier,
+                                                           Node<A, S> leaf,
+                                                           boolean direction) {
+    if (!exploredStates.contains(leaf.state())) {
+      thisFrontier.add(leaf);
+    }
+    if (exploredStates.contains(leaf.state())) {
+      return null;
+    } else {
+      exploredStates.add(leaf.state());
+    }
+    return meetingNodes(leaf, otherFrontier, direction);
   }
 
   private Node<A, S> newLeafNode(Problem<A, S> problem, Node<A, S>
@@ -222,25 +237,11 @@ public class BiDirectionalBreadthFirstSearch<A, S> {
     return new ConcurrentLinkedQueue<>();
   }
 
-  private Set<S> concurrentStateSet() {
-    return new ConcurrentSkipListSet<S>() {
-      @Override
-      public Comparator<? super S> comparator() {
-        return (Comparator<S>) (o1, o2) -> o1.toString().compareTo(o2.toString());
-      }
-    };
-  }
 
-  public List<A> failure() {
-    return searchController.failure();
-  }
-
-  public List<A> solution(Node<A, S> child) {
-    return searchController.solution(child);
-  }
-
-  public void register(StateActionTimeLine<String, String> timeLine) {
-    history = timeLine;
+  public static class Strategy {
+    public static final String BFS = "breadth first search";
+    public static final String UCS = "uniform cost search";
+    public static String ASTAR = "A*";
   }
 
   private class SearchDirection {
